backend/compositeservices/confirmStatus/requirements.txt
flask
flask-cors
requests
pika
python-dotenv
